+++ src/commands/hld/reconcile.test.ts
@@ -16,7 +16,6 @@ import {
   IReconcileDependencies,
   reconcileHld,
   testAndGetAbsPath,
-  tryGetGitOrigin,
   validateInputs
 } from "./reconcile";
 import * as reconcile from "./reconcile";
@@ -101,31 +100,6 @@ describe("checkForFabrikate", () => {
     expect(() => {
       checkForFabrikate(which);
     }).not.toThrow();
-  });
-});
-
-describe("tryGetGitOrigin", () => {
-  it("attempts to retrieve azdo git origin", async () => {
-    const originUrl = "http://github.com/repo/url";
-
-    (getAzdoOriginUrl as jest.Mock).mockReturnValue(originUrl);
-
-    const originUrlResponse = await tryGetGitOrigin();
-    expect(originUrlResponse).toEqual(originUrl);
-    expect(getAzdoOriginUrl).toHaveBeenCalledTimes(1);
-  });
-
-  it("attempts to retrieve git origin from using git cli", async () => {
-    const originUrl = "http://github.com/repo/url";
-    (getAzdoOriginUrl as jest.Mock).mockRejectedValue("some reason");
-
-    (getOriginUrl as jest.Mock).mockReturnValue(originUrl);
-
-    const originUrlResponse = await tryGetGitOrigin();
-
-    expect(originUrlResponse).toEqual(originUrl);
-    expect(getAzdoOriginUrl).toHaveBeenCalledTimes(1);
-    expect(getOriginUrl).toHaveBeenCalledTimes(1);
   });
 });
